updated: split lifecycle methods
