+++ lib/browser/common/util/misc/index.js
@@ -4,6 +4,7 @@ export { default as getPropDescriptor } from './get-prop-descriptor'
 export { default as startsWith } from './starts-with'
 export { default as uid } from './uid'
 export { default as defineProperty } from './define'
+export { default as objectCreate } from './object-create'
 export { default as extend } from './extend'
 export { default as toCamel } from './to-camel'
 export { default as warn } from './warn'
