+++ lib/browser/common/util/misc/object-create.js
@@ -1,0 +1,6 @@
+/**
+ * Alias for Object.create
+ */
+export default function create(src) {
+  return Object.create(src)
+}
