+++ lib/browser/common/util/tags/index.js
@@ -12,3 +12,5 @@ export { default as getImmediateCustomParentTag } from './get-immediate-custom-p
 export { default as makeVirtual } from './make-virtual'
 export { default as moveVirtual } from './move-virtual'
 export { default as unmountAll } from './unmount-all'
+export { default as createIfDirective } from './create-if-directive'
+export { default as createRefDirective } from './create-ref-directive'
