+++ lib/browser/common/util/tags/mount.js
@@ -4,6 +4,7 @@ import {
 } from './../../global-variables'
 import extend from '../misc/extend'
 import contains from '../misc/contains'
+import create from '../misc/object-create'
 import createTag from './../../../tag/tag'
 
 /**
@@ -17,7 +18,7 @@ import createTag from './../../../tag/tag'
 export default function mount(root, tagName, opts, ctx) {
   const impl = __TAG_IMPL[tagName]
   const implClass = __TAG_IMPL[tagName].class
-  const context = ctx || (implClass ? Object.create(implClass.prototype) : {})
+  const context = ctx || (implClass ? create(implClass.prototype) : {})
   // cache the inner HTML to fix #855
   const innerHTML = root._innerHTML = root._innerHTML || root.innerHTML
   const conf = extend({ root, opts, context }, { parent: opts ? opts.parent : null })
