+++ lib/browser/common/util/tags/update-options.js
@@ -1,0 +1,29 @@
+import each from '../misc/each'
+import { ATTRS_PREFIX } from './../../global-variables'
+import toCamel from '../misc/to-camel'
+import { updateExpression } from '../../../tag/update'
+import inheritParentProps from './inherit-parent-properties'
+
+/**
+ * We need to update opts for this tag. That requires updating the expressions
+ * in any attributes on the tag, and then copying the result onto opts.
+ * @this Tag
+ * @param   {Boolean} isLoop - is it a loop tag?
+ * @param   { Tag }  parent - parent tag node
+ * @param   { Boolean }  isAnonymous - is it a tag without any impl? (a tag not registered)
+ * @param   { Object }  opts - tag options
+ * @param   { Array }  instAttrs - tag attributes array
+ */
+export default function updateOpts(isLoop, parent, isAnonymous, opts, instAttrs) {
+  // isAnonymous `each` tags treat `dom` and `root` differently. In this case
+  // (and only this case) we don't need to do updateOpts, because the regular parse
+  // will update those attrs. Plus, isAnonymous tags don't need opts anyway
+  if (isLoop && isAnonymous) return
+  const ctx = isLoop ? inheritParentProps.call(this) : parent || this
+
+  each(instAttrs, (attr) => {
+    if (attr.expr) updateExpression.call(ctx, attr.expr)
+    // normalize the attribute names
+    opts[toCamel(attr.name).replace(ATTRS_PREFIX, '')] = attr.expr ? attr.expr.value : attr.value
+  })
+}
