+++ lib/browser/tag/each.js
@@ -15,8 +15,10 @@ import getAttribute from './../common/util/dom/get-attribute'
 import createPlaceholder from './../common/util/dom/create-placeholder'
 import safeInsert from './../common/util/dom/safe-insert'
 import createFragment from './../common/util/dom/create-fragment'
+
 import each from './../common/util/misc/each'
 import contains from './../common/util/misc/contains'
+import create from './../common/util/misc/object-create'
 import extend from './../common/util/misc/extend'
 
 import moveChild from './../common/util/tags/move-child'
@@ -45,7 +47,7 @@ import { tmpl } from 'riot-tmpl'
  *
  */
 function mkitem(expr, key, val, base) {
-  const item = base ? Object.create(base) : {}
+  const item = base ? create(base) : {}
   item[expr.key] = key
   if (expr.pos) item[expr.pos] = val
   return item
@@ -214,7 +216,7 @@ export default function _each(dom, parent, expr) {
         if (expr.key && !isObject)
           return !!tmpl(ifExpr, mkitem(expr, item, i, parent))
 
-        return !!tmpl(ifExpr, extend(Object.create(parent), item))
+        return !!tmpl(ifExpr, extend(create(parent), item))
       })
     }
