+++ lib/browser/tag/parse.js
@@ -1,5 +1,3 @@
-import IfExpr from './if'
-import RefExpr from './ref'
 import _each from './each'
 import { tmpl } from 'riot-tmpl'
 import {
@@ -9,6 +7,9 @@ import {
   REF_DIRECTIVES,
   YIELD_TAG
 } from './../common/global-variables'
+
+import createRefDirective from './../common/util/tags/create-ref-directive'
+import createIfDirective from './../common/util/tags/create-if-directive'
 import isBoolAttr from './../common/util/checks/is-boolean-attribute'
 import walkNodes from './../common/util/dom/walk-nodes'
 import getAttribute from './../common/util/dom/get-attribute'
@@ -55,7 +56,7 @@ export function parseExpressions(root, mustIncludeRoot) {
     // if-attrs become the new parent. Any following expressions (either on the current
     // element, or below it) become children of this expression.
     if (attr = getAttribute(dom, CONDITIONAL_DIRECTIVE)) {
-      expressions.push(Object.create(IfExpr).init(dom, this, attr))
+      expressions.push(createIfDirective(dom, this, attr))
       return false
     }
 
@@ -139,7 +140,7 @@ export function parseAttributes(dom, attrs, fn) {
     let expr
 
     if (contains(REF_DIRECTIVES, name) && dom.tagName.toLowerCase() !== YIELD_TAG) {
-      expr =  Object.create(RefExpr).init(dom, this, name, attr.value)
+      expr =  createRefDirective(dom, this, name, attr.value)
     } else if (tmpl.hasExpr(attr.value)) {
       expr = {dom, expr: attr.value, attr: name, bool}
     }
